Move FunctionDispatchItem into detail namespace
