+++ include/eventkit/dispatch/detail/DispatchQueue-inl.h
@@ -7,14 +7,14 @@
 
 #include <eventkit/dispatch/DispatchQueue.h>
 #include <eventkit/common/Allocator.h>
-#include <eventkit/dispatch/FunctionDispatchItem.h>
+#include <eventkit/dispatch/detail/FunctionDispatchItem.h>
 
 namespace ek {
 namespace dispatch {
 
 template <typename F>
 void DispatchQueue::dispatchAsync(ek::common::Allocator* pA, F&& function) {
-    dispatchItemAsync(ek::common::make_intrusive<FunctionDispatchItem<std::decay_t<F>>>(pA, pA, std::forward<F>(function)));
+    dispatchItemAsync(ek::common::make_intrusive<detail::FunctionDispatchItem<std::decay_t<F>>>(pA, pA, std::forward<F>(function)));
 }
 
 }
